separate out class into its own file
